backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35

# ULID
python-ulid==2.7.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Linting
ruff==0.6.8
